Fix bug that caused premature exit of launcher
Observed when DeepHyper was running on a login node and was dynamically
submitting ~4 second jobs to a single worker from a Balsam launcher with
--job-mode=mpi.

Having a single Balsam worker with DeepHyper's blocking behavior (would
not submit another Balsam job until it receives the previous result)
meant that there was never a job with state PREPROCESSED in the Balsam
queue when check_exit() was called ---> exit_counter was never reset to
zero.

In between DeepHyper submissions, the exit_counter would be incremented
as no job would be running.